+++ README.md
@@ -1,2 +1,4 @@
 # MasterSchedule
 School course master schedule organizer and validator
+
+FILES BEGINNING WITH F ARE FLASK EXAMPLES. THEY WERE RENAMED TO AVOID CONFLICT WITH OTHER FILES.
