+++ src/allsolutions.py
