+++ src/static/fscript.js
