+++ src/static/styles.css
@@ -1,0 +1,64 @@
+:root {
+    font-family: sans-serif;
+    --primaryColor: #800000;
+    --secondaryColor: #FFD700;
+    --teritaryColor: #000000;
+    --backgroundColor: #F5F5DC;
+    scroll-behavior: smooth;
+}
+
+header {
+    text-align: center;
+    background-color: var(--primaryColor);
+    color: var(--secondaryColor);
+    padding: 15px;
+    font-size: 20px;
+}
+
+body {
+    background-color: var(--backgroundColor);
+}
+
+h2 {
+    text-align: center;
+    color: red;
+    text-decoration: underline;
+}
+
+h3 {
+    text-align: center;
+}
+
+p {
+    text-align: center;
+}
+
+footer {
+    text-align: center;
+    background-color: var(--primaryColor);
+    color: var(--secondaryColor);
+    padding: 15px;
+    font-size: 20px;
+}
+
+a {
+    text-decoration: underline;
+    color: var(--secondaryColor);
+}
+
+a:hover {
+    text-decoration: underline;
+    color: var(--teritaryColor)
+}
+
+button {
+    background-color: var(--secondaryColor);
+    color: var(--primaryColor);
+    border: 3px solid var(--teritaryColor);
+    font-size: 20px;
+}
+
+button:hover {
+    background-color: var(--primaryColor);
+    color: var(--secondaryColor);
+}
